refactor(cli): separate stack into own class
